tinylog: use std::unique_ptr

+++ libs/tinylog/tinylog/Config.h
@@ -3,4 +3,4 @@
 #include <cstdint>
 
 const int32_t TIME_OUT_SECOND = 3;
-const uint64_t BUFFER_SIZE = 24 * 1024;
+const uint32_t BUFFER_SIZE = 24 * 1024;
